Re-Organization & Swipe menu
Re-Organization of the game scripts into appropriate folders.
Script creation for a swipe menu functionality.

+++ GameMechanics/CountDownSystem.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 
 public class CountDownSystem : MonoBehaviour
 {
-    public float maxTime;
-    public float timeLeft;
+    public double maxTime;
+    public double timeLeft;
     public GameOverSystem gameOverSystem;
     public AudioSource weeds;
