+++ SaveSystems/GameMaster.cs
@@ -23,7 +23,7 @@ public class GameMaster : MonoBehaviour
     public int totalPoints;
     public int totalMoney;
     public int totalStars = 9;
-    public List<float> bestTimes = new List<float>();
+    public List<double> bestTimes = new List<double>();
     public List<int> bestScores = new List<int>();
     public List<int> bestMorales = new List<int>();
     public List<int> bestStars = new List<int>();
